refactor: :truck: moved the util folder
