+++ services/authService.ts
@@ -1,4 +1,4 @@
-import { clearTokens, getToken, saveToken } from "@/app/utils/secureStorage";
+import { clearTokens, getToken, saveToken } from "@/utils/secureStorage";
 import axios, { isAxiosError } from "axios";
 import { Platform } from "react-native";
